requirements.txt
python>=3.7
dash>=1.4.1
numpy>=1.9
Pillow>=7.2
pyserial>=3.4
flask>=1.1.2
setuptools>=49.2.0
# vimba>=4.1 should also be installed. Download it from the alliedvision site and follow instructions.